11th Commit
Separation of Concerns - Refactoring 1

The FormComponent
All the domain logic has been removed from the FormComponent. The FormComponent subscribes only to the ChangeService – getFormFields observable, for obtaining a formFields array. The DataService, as well as the ItemFomFields constant are totally unknown to the FormComponent. Only the interface IFormField is needed.

The GetDataComponent
The GetDataComponent still accesses the DataService to initialize the Categories array and to get an Item whenever we click on the “Get It” button, and updates the ChangeService accordingly. Categories are needed once, but they are needed early, for setting initial options array, so the GetDataComponent updates the Categories in its ngOnInit life-cycle hook.

The ChangeService
The ChangeService becomes the “bridge” between the domain model (e.g.: DataService) and the metadata that describes the business object model (IFormField, ItemFomFields). The updateOptions() and the updateFormFieldsInitialValues() methods have been moved in it as private methods, from the FormComponent.  The ChangeService now, also uses the getFormFields$$ BehaviorSubject, as observable via the getFormFields() public method, to expose the formFields array. This is actually, the only one “contact” point with the FormComponent, which subscribes to it (in its ngOnInit() life-cycle hook method) and it is being kept updated with changes in the formFields array. Thus, the formFields array can be considered as a dynamic formFields array of objects.
itemIdChange$$ and categories$$ are still used to be exposed as observables in the getItem() and the getCategories() public methods respectively.

Other refactoring