13th Commit
Refactoring 1 continued
The ChangeService renamed to ‘ItemsFormFieldsService’
All domain logic has been removed from the GetDataComponent. Now, the GetDataComponent, deals only with the ItemsFormFieldsService, for sending it just the itemId, whenever the user clicks on the “Get It” button. It does it, via the setItemId() method of the ItemsFormFieldsService.
The DataService is now injected in the ItemsFormFieldsService. So, the ItemsFormFieldsService is solo responsible to deal with it. It uses it to fetch categories just once in its constructor. In the new public method setItemId(), it also accesses the DataService to obtain an IItem object, each time the id (itemId) is updated. Then it calls the (private) setItem() method to update form-fields initial values and “push” it also, in the   itemIdChange$$ sunject.